Struct binjs::ast::annotation::Context

pub struct Context<'a, T> where
    ContextContents<'a, T>: Dispose,  { /* fields omitted */ }

Storage for information collected while annotating an AST.

Methods

impl<'a, T> Context<'a, T> where
T: Default,
ContextContents<'a, T>: Dispose,

fn new(syntax: &'a Syntax) -> Self

fn enter_field(&mut self, field: &str) -> Result<Self, ASTError>

fn enter_obj(&mut self, kind: &str) -> Result<Self, ASTError>

fn kind_str(&self) -> &str

fn contents(&self) -> Ref<ContextContents<'a, T>>

fn parent(&self) -> Option<Rc<RefCell<ContextContents<'a, T>>>>

impl<'a> Context<'a, RefContents>

fn add_direct_eval(&mut self)

fn add_free_name(&mut self, name: &str)

fn add_bound_name(&mut self, name: &str)

fn is_bound(&self, name: &str) -> bool

Check whether the name is bound somewhere on the stack.

fn store(&self, object: &mut Map<String, JSON>)

fn load(&mut self, object: &Map<String, JSON>)

impl<'a> Context<'a, DeclContents>

fn add_var_name(&mut self, name: &str)

fn add_lex_name(&mut self, name: &str)

fn is_lex_bound(&self, name: &str) -> bool

Check whether the name is lexically bound somewhere on the stack.

fn scope_kind(&self) -> ScopeKind

fn set_scope_kind(&mut self, kind: ScopeKind)

fn clear_var_names(&mut self)

fn clear_lex_names(&mut self)

fn store(&self, object: &mut Map<String, JSON>)

Store scope information in a node.

This makes sense only if the node implements binjs, otherwise all data will be stripped while encoding.