Module binjs::ast::grammar

Definition of a grammar, tools for describing the grammar. Grammars for specifying an AST that this tool can manipulate.

One of the most important specificities of BinJS is that it is a format designed for a living language. Critically, a BinJS file that can be parsed today must still be parsed after any number of evolutions of the language, without having to write a distinct parser per version of the language.

To ensure this, we define not only the AST of EcmaScript as the language is today, but the mechanism to define and evolve this AST, without incurring breaking changes in BinJS.

Structs

Field
Representation of a field in an interface.

FieldName
The name of a field in an interface.

Interface
An interface, once compiled through SyntaxBuilder::as_syntax.

InterfaceDeclaration

NodeName
The name of an interface or enum.

Obj
Representation of an object, i.e. a set of fields.

StringEnum
An enumeration of strings.

Syntax
Immutable representation of the syntax.

SyntaxBuilder
A data structure used to progressively construct the Syntax.

SyntaxOptions
Informations passed during the creation of a Syntax object.

Type

TypeSum
An enumeration of interfaces.

Enums

ASTError

NamedType

TypeSpec
A type, typically that of a field.