Struct binjs::ast::grammar::Obj

pub struct Obj { /* fields omitted */ }

Representation of an object, i.e. a set of fields.

Field order is not specified, but is expected to remain stable during encoding operations and during decoding operations. Note in particular that the order may change between encoding and decoding.

Methods

impl Obj

pub fn new() -> Self

Create a new empty structure

pub fn fields<'a>(&'a self) -> &'a [Field]

A list of the fields in the structure.

pub fn field<'a>(&'a self, name: &FieldName) -> Option<&'a Field>

Fetch a specific field in the structure

pub fn with_own_field(self, field: Field) -> Self

pub fn with_field(self, name: &FieldName, type_: Type) -> Self

Extend a structure with a field.

pub fn with_field_doc(self, name: &FieldName, type_: Type, doc: &str) -> Self