Struct binjs::ast::grammar::Syntax

pub struct Syntax { /* fields omitted */ }

Immutable representation of the syntax.

Methods

impl Syntax

pub fn get_ancestors_by_name(&self, name: &NodeName) -> Option<&[NodeName]>

Return all the ancestors of an interface, including itself.

pub fn get_interface_by_kind(&self, kind: &Kind) -> Option<&Interface>

pub fn get_interface_by_name(&self, name: &NodeName) -> Option<&Interface>

pub fn get_enum_by_name(&self, name: &NodeName) -> Option<Ref<Enum>>

pub fn get_kind(&self, name: &str) -> Option<&Kind>

pub fn get_field_name(&self, name: &str) -> Option<&FieldName>

pub fn get_node_name(&self, name: &str) -> Option<&NodeName>

pub fn get_root(&self) -> &Interface

The starting point for parsing.

pub fn validate(&self, a: &JSON) -> Result<(), ASTError>

Ensure that a value is an inhabitant of the grammar.

pub fn validate_from(&self, a: &JSON, type_: &Type) -> Result<(), ASTError>

Ensure that a value is an inhabitant of the grammar.

pub fn annotate(&self, tree: &mut JSON) -> Result<(), ASTError>

pub fn compare(&self, a: &JSON, b: &JSON) -> Result<bool, ASTError>

Compare two ASTs, restricting comparison to the items that appear in the grammar.

This method assumes that both items are full ASTs.

pub fn compare_from(
&self,
left: &JSON,
right: &JSON,
inhabit: &Type
) -> Result<bool, ASTError>

Compare two ASTs, restricting comparison to the items that appear in the grammar.

pub fn has_ancestor_in(
&self,
interface: &Interface,
one_of: &[NodeName]
) -> bool

pub fn random<T: Rng>(&self, rng: &mut T, depth_limit: isize) -> JSON

Generate a random AST matching the grammar.

depth_limit is used as hint to control the depth of the tree

pub fn pretty(&self, indent: &str) -> String