Struct binjs::ast::grammar::SyntaxBuilder

pub struct SyntaxBuilder { /* fields omitted */ }

A data structure used to progressively construct the Syntax.

Methods

impl SyntaxBuilder

fn new() -> Self

fn node_name(&mut self, name: &str) -> NodeName

Return an NodeName for a name. Equality comparison on NodeName can be performed by checking physical equality.

fn field_name(&mut self, name: &str) -> FieldName

fn kind_name(&mut self, name: &str) -> Kind

fn add_kinded_interface(
&mut self,
name: &NodeName
) -> Option<RefMut<InterfaceDeclaration>>

Add an interface with a kind identical to its name.

fn add_virtual_interface(
&mut self,
name: &NodeName
) -> Option<RefMut<InterfaceDeclaration>>

Add a virtual interface, i.e. one that doesn't have a kind, i.e. one that does not have immediate inhabitants. Super-interfaces or sub-interfaces with a kind may have inhabitants.

fn add_enum(&mut self, name: &NodeName) -> Option<RefMut<Enum>>

Add a named enumeration.

fn into_syntax<'a>(self, options: SyntaxOptions<'a>) -> Syntax

Generate the graph.