Struct binjs::ast::grammar::SyntaxOptions

pub struct SyntaxOptions<'a> {
    pub null: &'a Kind,
    pub root: &'a NodeName,
    pub annotator: Box<Annotator>,
}

Informations passed during the creation of a Syntax object.

Fields
null: &'a Kind

The kind of the special node used to encode null AST nodes (NOT null literals).

root: &'a NodeName

The name of the node used to start encoding.

annotator: Box<Annotator>