Struct binjs::ast::grammar::Type

pub struct Type {
    pub spec: TypeSpec,
    pub defaults_to: Option<JSON>,
}

Fields
spec: TypeSpec defaults_to: Option<JSON>

If the value is not specified, it defaults to... (None if the value MUST be specified)

Methods

impl Type

pub fn with_default(&mut self, default: JSON) -> &mut Self

pub fn with_named_sum(&mut self, names: &[&NodeName]) -> &mut Self

pub fn with_spec(&mut self, spec: TypeSpec) -> &mut Self

pub fn spec(&self) -> &TypeSpec

pub fn default(&self) -> Option<&JSON>

pub fn named(name: &NodeName) -> TypeSpec

Shorthand constructors.

pub fn sum(types: &[TypeSpec]) -> TypeSpec

pub fn string() -> TypeSpec

pub fn number() -> TypeSpec

pub fn bool() -> TypeSpec

pub fn random<T: Rng>(
&self,
syntax: &Syntax,
rng: &mut T,
depth_limit: isize
) -> JSON

pub fn pretty(&self, prefix: &str, indent: &str) -> String

pub fn validate(&self, syntax: &Syntax, a: &JSON) -> Result<(), ASTError>

Ensure that a value is an inhabitant of the grammar.

pub fn compare(
&self,
syntax: &Syntax,
left: &JSON,
right: &JSON
) -> Result<bool, ASTError>

Compare two ASTs, restricting comparison to the items that appear in the grammar.