Module binjs::generic

Working with a generic (i.e. JSON-based) representation of the JavaScript AST.

Modules

ast

es6
Generic instance of Spec representing the es6 AST.

mru
Naive implementation of a MRU numbering scheme.

pick
Generating random ASTs (for fuzzing purposes).

util
Walkers, comparisons.

Macros

const_with_str
From a set of macro definitions, derive a module with a set of matching &'static str.

shared_string

Structs

F64
A container for f64 values that implements an arbitrary total order, equality relation, hash.

FieldName
A field name of the grammar.

IdentifierName
An identifier, inside the grammar.

InterfaceName
An interface of the grammar.

Offset

PropertyKey
A property, inside the grammar.

Enums

SharedString
An implementation of strings that may easily be shared without copies.

VisitMe

Traits

Node
The root type for nodes in the AST.

Type Definitions

IOPath
The type of paths used most commonly throughout our code.

IOPathItem
The type of path items used most commonly throughout our code.

Identifier