Trait binjs::generic::ast::Node

pub trait Node: Downcast {
    fn name(&self) -> &'static str;

    fn scoped_dictionary(&self) -> Option<&SharedString> { ... }
}

The root type for nodes in the AST.

Required methods

fn name(&self) -> &'static str

Return the name of the current node.

Loading content...
Provided methods

fn scoped_dictionary(&self) -> Option<&SharedString>

If this node should cause a dictionary change, return the name of the new dictionary to use. Otherwise, None.

Loading content...
Methods

impl dyn Node + 'static

pub fn is<__T>(&self) -> bool where
__T: Node,

Returns true if the trait object wraps an object of type __T.

pub fn downcast<__T>(
self: Box<dyn Node + 'static>
) -> Result<Box<__T>, Box<dyn Node + 'static>> where
__T: Node,

Returns a boxed object from a boxed trait object if the underlying object is of type __T. Returns the original boxed trait if it isn't.

pub fn downcast_rc<__T>(
self: Rc<dyn Node + 'static>
) -> Result<Rc<__T>, Rc<dyn Node + 'static>> where
__T: Node,

Returns an Rc-ed object from an Rc-ed trait object if the underlying object is of type __T. Returns the original Rc-ed trait if it isn't.

pub fn downcast_ref<__T>(&self) -> Option<&__T> where
__T: Node,

Returns a reference to the object within the trait object if it is of type __T, or None if it isn't.

pub fn downcast_mut<__T>(&mut self) -> Option<&mut __T> where
__T: Node,

Returns a mutable reference to the object within the trait object if it is of type __T, or None if it isn't.