Module binjs::io

Language-agnostic and representation-agnostic primitives for tokenization and detokenization of ASTs, including (de)compression.

Weakly-typed constructs (using a dynamic language specification and a JSON AST) based on these primitives are in module generic::io. Strongly-typed constructs (using a strongly-typed AST) based on these primitives will appear in modules specialized::*::io.

Modules

bytes
Byte-level utilities for writing token readers/writers. Tools for manipulating byte-level data.

entropy
An encoding using entropy coding. A format in which we adopt an arithmetic encoding per nature.

io
Definition of TokenReader/TokenWriter traits. Mid-level API used to write individual components of an AST tree to a stream of tokens.

multipart
An optimization of TokenReader/TokenWriter, designed to minimize the size of the file. A multipart format, in which each part can be compressed independently.

simple
A simple implementation of TokenReader/TokenWriter, designed specifically to help debug implementations of grammar encoders/decoders. Minimal implementation of encoding/decoding to binary. Used for testing purposes. Not meant to be included in release builds.

xml
A trivial exporter to xml.

Macros

print_file_structure
Prints the structural interpretation of the data read after the last time this macro is called.

Structs

CompressionTarget
Instructions for a single section (grammar, strings, tree, ...)

Enums

DictionaryPlacement
A strategy for placing the dictionary.

Format
All the formats available for encoding/decoding.

TokenReaderError

TokenWriterError

Traits

FormatProvider
Command-line management for a format