Module binjs::io::binjs_json::read

A JSON reader with extra support for JSON->BinAST converter. The input JSON file is supposed to be the output of the JSON writer, filtered with some command to generate an invalid content. See mod.rs for the detail of the filter. Some methods implemented on Decoder are for binjs_convert_from_json command.

Structs

Decoder

Enums

NextType
The next item in the tree, designates which method to call on the decoder.

Type Definitions

PathItem