Module binjs::io::entropy::rw

Tools shared by read and write. Constants and data structures shared between entropy read and entropy write.

Structs

PreludeStreams
Prelude data.

TableRefStreamState
State of a input/output stream of TableRef.

Constants

FORMAT_BROTLI
Indicating that the stream is compressed with format "brotli".

FORMAT_ENTROPY_0
Indicating that the stream is compressed with format "entropy 0.2".

GLOBAL_HEADER_START
The magic number at the start of the file.

NAME_MAX_LEN
A stack-allocated buffer for names of sections and streams.

SECTION_CONTENT
The content section, containing the streams of indices of user-extensible values, encoded e.g. using brotli.

SECTION_CONTENT_WITHOUT_BRACKETS

SECTION_MAIN
The main section, containing the entropy-encoded stream.

SECTION_MAIN_WITHOUT_BRACKETS

SECTION_PRELUDE
The prelude section, containing the dictionary extensions, encoded using e.g. brotli.

Type Definitions

NameData