Module binjs::meta::spec

Manipulating the specifications of the language. Definition of a spec for a version (or subset) of JavaScript.

Structs

Field
Representation of a field in an interface.

FieldName
The name of a field in an interface.

Interface
Representation of an interface in a grammar declaration.

InterfaceDeclaration

IsNullable

NodeName
The name of an interface or enum.

Obj
Representation of an object, i.e. a set of fields.

Spec
Immutable representation of the spec.

SpecBuilder
A data structure used to progressively construct the Spec.

SpecOptions
Informations passed during the creation of a Spec object.

StringEnum
An enumeration of strings.

Type

TypeSum
An enumeration of interfaces.

Enums

Laziness
Whether an attribute is eager or lazy.

NamedType

Primitive

TypeSpec
The contents of a type, typically that of a field.

Traits

HasInterfaces

ToStr