Trait binjs::token::io::Guard

pub trait Guard {
    type Error;
    fn done(self) -> Result<(), Self::Error>;
}

A guard used to ensure that some subset of the input stream was read properly.

Associated Types

type Error

The type of errors returned by the guard. This is typically TokenReader::Error.

Required Methods

fn done(self) -> Result<(), Self::Error>

Ensure that the subset of the input stream was read properly.