Trait binjs::token::io::TokenReader

pub trait TokenReader where
    Self::Error: Debug,
    Self::ListGuard: Guard<Error = Self::Error>,
    Self::TaggedGuard: Guard<Error = Self::Error>,
    Self::UntaggedGuard: Guard<Error = Self::Error>,
    Self: Sized, {
    type Error;
    type ListGuard;
    type TaggedGuard;
    type UntaggedGuard;
    fn poison(&mut self);

    fn string(&mut self) -> Result<Option<String>, Self::Error>;

    fn float(&mut self) -> Result<Option<f64>, Self::Error>;

    fn bool(&mut self) -> Result<Option<bool>, Self::Error>;

    fn list(&mut self) -> Result<(u32, Self::ListGuard), Self::Error>;

    fn tagged_tuple(
        &mut self
    ) -> Result<(String, Rc<Box<[Field]>>, Self::TaggedGuard), Self::Error>;

    fn untagged_tuple(&mut self) -> Result<Self::UntaggedGuard, Self::Error>;
}

An API for reading tokens.

Note that a TokenReader by itself cannot determine the nature of the following token. Rather, the driver of the TokenReader must be able to deduce the nature of the following token from what it has previously read.

Associated Types

type Error

An error returned by the extractor.

Errors are not recoverable within a TokenReader.

For instance, if attempting to read with string() fails, any further attempt to use the TokenReader or any of its parents will also raise an error.

type ListGuard

A guard, used to make sure that the consumer has properly read a list.

See the documentation of self.list.

type TaggedGuard

A guard, used to make sure that the consumer has properly read a tagged tuple.

See the documentation of self.tagged_tuple.

type UntaggedGuard

A guard, used to make sure that the consumer has properly read an untagged tuple.

See the documentation of self.untagged_tuple.

Required Methods

fn poison(&mut self)

Poison the reader, ensuring that it will never be used for reading again.

fn string(&mut self) -> Result<Option<String>, Self::Error>

Read a single UTF-8 string.

The returned string MUST be valid UTF-8.

fn float(&mut self) -> Result<Option<f64>, Self::Error>

Read a single f64. Note that all numbers are f64.

fn bool(&mut self) -> Result<Option<bool>, Self::Error>

Read a single bool.

fn list(&mut self) -> Result<(u32, Self::ListGuard), Self::Error>

Start reading a list.

Returns a guard for that list and the number of elements in the list. Once the list is read entirely, callers MUST call guard.done() to ensure that the list was properly read (in particular that all bytes were consumed). In most implementations, failure to do so will raise an assertion.

fn tagged_tuple(
&mut self
) -> Result<(String, Rc<Box<[Field]>>, Self::TaggedGuard), Self::Error>

Start reading a tagged tuple. If the stream was encoded properly, the tag is attached to an ordered tuple of fields that may be extracted in order.

Returns the tag name, the ordered array of fields in which the contents must be read, and a guard for that tuple. Once the tuple is read entirely, callers MUST call guard.done() to ensure that the tuple was properly read (in particular that all bytes were consumed). In most implementations, failure to do so will raise an assertion.

fn untagged_tuple(&mut self) -> Result<Self::UntaggedGuard, Self::Error>

Start reading an untagged tuple.

Once the tuple is read entirely, callers MUST call guard.done() to ensure that the tuple was properly read (in particular that all bytes were consumed). In most implementations, failure to do so will raise an assertion.