Module binjs::util

Misc utilities.

Structs

PoisonLock
A structure used to make sure that every access to a state goes through try and that any error poisons the state.

Traits

Dispose

JSONGetter
Utilities to simplify dealing with JSON.

Pos
An object (typically a reader) that knows its position and size.

ReadConst
An extension of Read that knows how to check that the following few bytes match some value.

Functions

f64_of

get_temporary_file

strip
Strip a tree from meaningless information (location information, comments, ...)

type_of
Utility: return a string describing a JSON value (e.g. "Object", "Bool", ...)