Struct binjs::util::PoisonLock

pub struct PoisonLock<S> { /* fields omitted */ }

A structure used to make sure that every access to a state goes through try and that any error poisons the state.

Methods

impl<S> PoisonLock<S>

fn new(state: S) -> Self

fn try<T, E, F>(&mut self, f: F) -> Result<T, E> where
F: FnOnce(&mut S) -> Result<T, E>,

Access the state for an operation.

If the operation fails, the state is poisoned.

Panics

If the ReaderState is poisoned.

fn poison(&mut self)

fn is_poisoned(&self) -> bool