Struct binjs::ast::annotation::Context

pub struct Context<'a, T> where
    ContextContents<'a, T>: Dispose,  { /* fields omitted */ }

Storage for information collected while annotating an AST.

Methods

impl<'a, T> Context<'a, T> where
T: Default,
ContextContents<'a, T>: Dispose,

pub fn new(syntax: &'a Syntax) -> Self

pub fn lex_scope(&self) -> Option<Rc<RefCell<ContextContents<'a, T>>>>

pub fn fun_scope(&self) -> Option<Rc<RefCell<ContextContents<'a, T>>>>

pub fn var_scope(&self) -> Option<Rc<RefCell<ContextContents<'a, T>>>>

pub fn use_as_fun_scope(&mut self)

pub fn use_as_lex_scope(&mut self)

pub fn use_as_var_scope(&mut self)

pub fn enter_field(&mut self, field: &str) -> Result<Self, ASTError>

pub fn enter_obj(&mut self, kind: &str) -> Result<Self, ASTError>

pub fn kind_str(&self) -> &str

pub fn contents(&self) -> Ref<ContextContents<'a, T>>

pub fn parent(&self) -> Option<Rc<RefCell<ContextContents<'a, T>>>>

impl<'a> Context<'a, RefContents>

pub fn add_direct_eval(&mut self)

pub fn add_free_name(&mut self, name: &str)

pub fn add_binding(&mut self, name: &str)

pub fn is_bound(&self, name: &str) -> bool

Check whether the name is bound somewhere on the stack.

pub fn store(&self, parent: &mut Object)

pub fn load(&mut self, parent: &Object)

impl<'a> Context<'a, DeclContents>

pub fn add_var_name(&mut self, name: &str)

pub fn add_let_name(&mut self, name: &str)

pub fn add_const_name(&mut self, name: &str)

pub fn is_lex_bound(&self, name: &str) -> bool

Check whether the name is lexically bound somewhere on the stack.

pub fn scope_kind(&self) -> ScopeKind

pub fn set_scope_kind(&mut self, kind: ScopeKind)

pub fn promote_unknown_names_to_var(&mut self)

pub fn clear_special_unknown_names(&mut self, names: &[&str])

pub fn clear_var_names(&mut self)

pub fn clear_lex_names(&mut self)

pub fn uses_strict(&self) -> bool

pub fn set_uses_strict(&mut self, value: bool)

pub fn store(&self, parent: &mut Object)

Store scope information in a field "BINJS:Scope" of the node.