Module binjs::ast::grammar

Tools for describing the grammar. Grammars for specifying an AST that this tool can manipulate.

This abstracts away the concepts of ESTree.

Structs

Enum
Structure of an enum of strings.

Field

FieldName

Interface

InterfaceNode
An interface, with additional data computed during the call to SyntaxBuilder::as_syntax.

Kind

NodeName

Obj
Obj of an object-like value.

Syntax
Immutable representation of the syntax.

SyntaxBuilder
A data structure used to progressively construct the Syntax.

Enums

Type
A type, typically that of a field.